Fixes symlink() error during template install on Windows

+++ spec/overcommit/hook_context/pre_commit_spec.rb
@@ -154,7 +154,7 @@ describe Overcommit::HookContext::PreCommit do
       around do |example|
         repo do
           `git commit --allow-empty -m "Initial commit"`
-          FileUtils.touch 'some-file'
+          touch 'some-file'
           `git add some-file`
           `git commit -m "Add file"`
           `git mv some-file renamed-file`
@@ -198,19 +198,22 @@ describe Overcommit::HookContext::PreCommit do
       end
     end
 
-    context 'when a broken symlink is staged' do
-      around do |example|
-        repo do
-          Overcommit::Utils::FileUtils.symlink('non-existent-file', 'symlink')
-          `git add symlink`
-          example.run
+    # Git cannot track Windows symlinks
+    unless Overcommit::OS.windows?
+      context 'when a broken symlink is staged' do
+        around do |example|
+          repo do
+            Overcommit::Utils::FileUtils.symlink('non-existent-file', 'symlink')
+            `git add symlink`
+            example.run
+          end
         end
-      end
 
-      it 'does not attempt to update/restore the modification time of the file' do
-        File.should_not_receive(:mtime)
-        File.should_not_receive(:utime)
-        subject
+        it 'does not attempt to update/restore the modification time of the file' do
+          File.should_not_receive(:mtime)
+          File.should_not_receive(:utime)
+          subject
+        end
       end
     end
   end
@@ -410,7 +413,7 @@ describe Overcommit::HookContext::PreCommit do
 
     it 'does not include submodules' do
       submodule = repo do
-        FileUtils.touch 'foo'
+        touch 'foo'
         `git add foo`
         `git commit -m "Initial commit"`
       end
@@ -434,7 +437,7 @@ describe Overcommit::HookContext::PreCommit do
     context 'when files were added' do
       around do |example|
         repo do
-          FileUtils.touch('some-file')
+          touch('some-file')
           `git add some-file`
           example.run
         end
@@ -446,7 +449,7 @@ describe Overcommit::HookContext::PreCommit do
     context 'when files were modified' do
       around do |example|
         repo do
-          FileUtils.touch('some-file')
+          touch('some-file')
           `git add some-file`
           `git commit -m "Initial commit"`
           echo('Hello', 'some-file')
@@ -461,7 +464,7 @@ describe Overcommit::HookContext::PreCommit do
     context 'when files were deleted' do
       around do |example|
         repo do
-          FileUtils.touch('some-file')
+          touch('some-file')
           `git add some-file`
           `git commit -m "Initial commit"`
           `git rm some-file`
@@ -475,10 +478,10 @@ describe Overcommit::HookContext::PreCommit do
     context 'when amending last commit' do
       around do |example|
         repo do
-          FileUtils.touch('some-file')
+          touch('some-file')
           `git add some-file`
           `git commit -m "Initial commit"`
-          FileUtils.touch('other-file')
+          touch('other-file')
           `git add other-file`
           example.run
         end
@@ -495,7 +498,7 @@ describe Overcommit::HookContext::PreCommit do
       around do |example|
         repo do
           `git commit --allow-empty -m "Initial commit"`
-          FileUtils.touch 'some-file'
+          touch 'some-file'
           `git add some-file`
           `git commit -m "Add file"`
           `git mv some-file renamed-file`
@@ -512,51 +515,54 @@ describe Overcommit::HookContext::PreCommit do
       end
     end
 
-    context 'when changing a symlink to a directory during an amendment' do
-      around do |example|
-        repo do
-          `git commit --allow-empty -m "Initial commit"`
-          FileUtils.mkdir 'some-directory'
-          FileUtils.ln_s 'some-directory', 'some-symlink'
-          `git add some-symlink some-directory`
-          `git commit -m "Add file"`
-          `git rm some-symlink`
-          FileUtils.mkdir 'some-symlink'
-          FileUtils.touch File.join('some-symlink', 'another-file')
-          `git add some-symlink`
-          example.run
+    # Git cannot track Windows symlinks
+    unless Overcommit::OS.windows?
+      context 'when changing a symlink to a directory during an amendment' do
+        around do |example|
+          repo do
+            `git commit --allow-empty -m "Initial commit"`
+            FileUtils.mkdir 'some-directory'
+            symlink('some-directory', 'some-symlink')
+            `git add some-symlink some-directory`
+            `git commit -m "Add file"`
+            `git rm some-symlink`
+            FileUtils.mkdir 'some-symlink'
+            touch File.join('some-symlink', 'another-file')
+            `git add some-symlink`
+            example.run
+          end
+        end
+
+        before do
+          context.stub(:amendment?).and_return(true)
+        end
+
+        it 'does not include the directory in the list of modified files' do
+          subject.should_not include File.expand_path('some-symlink')
         end
       end
 
-      before do
-        context.stub(:amendment?).and_return(true)
-      end
-
-      it 'does not include the directory in the list of modified files' do
-        subject.should_not include File.expand_path('some-symlink')
-      end
-    end
-
-    context 'when breaking a symlink during an amendment' do
-      around do |example|
-        repo do
-          `git commit --allow-empty -m "Initial commit"`
-          FileUtils.mkdir 'some-directory'
-          FileUtils.touch File.join('some-directory', 'some-file')
-          FileUtils.ln_s 'some-directory', 'some-symlink'
-          `git add some-symlink some-directory`
-          `git commit -m "Add file"`
-          `git rm -rf some-directory`
-          example.run
+      context 'when breaking a symlink during an amendment' do
+        around do |example|
+          repo do
+            `git commit --allow-empty -m "Initial commit"`
+            FileUtils.mkdir 'some-directory'
+            touch File.join('some-directory', 'some-file')
+            symlink('some-directory', 'some-symlink')
+            `git add some-symlink some-directory`
+            `git commit -m "Add file"`
+            `git rm -rf some-directory`
+            example.run
+          end
         end
-      end
 
-      before do
-        context.stub(:amendment?).and_return(true)
-      end
+        before do
+          context.stub(:amendment?).and_return(true)
+        end
 
-      it 'still includes the broken symlink in the list of modified files' do
-        subject.should include File.expand_path('some-symlink')
+        it 'still includes the broken symlink in the list of modified files' do
+          subject.should include File.expand_path('some-symlink')
+        end
       end
     end
   end
